handle non-decodable filenames
we assume the scanned filesystem uses utf8, but sometimes some files are
encoded differently.
python's listdir returns bytes filenames as unicode decoded with
"surrogateescape" error handler, for example "\udce9" for a lone byte b"\xe9".
sqlite lib can't insert those codepoints and we can't seem to control its
encoding so we will insert those names as BLOB in sqlite, not TEXT.

Then we need to be able to handle TEXT and BLOB in decoyfs.