for performance, use a single sqlite database
When multithreaded, fuse opens a new thread for each VFS operation, which
means (in previous code) opening a new sqlite connection in each thread.
As a result, running a simple "tree" command was a lot faster in single-thread
mode than in multi-thread.

Since we're read-only, we can have a single connection and let sqlite3 lock
when appropriate (and force single-thread if sqlite3 is not able to lock
by itself (safety is 0 or 1)).

If performance is still not enough in the future, we could have a pool of
sqlite connections, with connections released to the pool when a thread
finishes (when a VFS operation finishes).